+++ android/engine/src/test/java/org/smartregister/fhircore/engine/ui/questionnaire/QuestionnaireViewModelTest.kt
@@ -35,6 +35,8 @@ import org.hl7.fhir.r4.model.Bundle
 import org.hl7.fhir.r4.model.CanonicalType
 import org.hl7.fhir.r4.model.CodeableConcept
 import org.hl7.fhir.r4.model.Coding
+import org.hl7.fhir.r4.model.Expression
+import org.hl7.fhir.r4.model.Extension
 import org.hl7.fhir.r4.model.Patient
 import org.hl7.fhir.r4.model.Questionnaire
 import org.hl7.fhir.r4.model.QuestionnaireResponse
@@ -130,19 +132,13 @@ class QuestionnaireViewModelTest : RobolectricTest() {
     )
 
     coVerify { defaultRepo.save(any()) }
-    coVerify { ResourceMapper.extract(any(), any(), any(), any()) }
+    coVerify { ResourceMapper.extract(any(), any(), any(), any(), any()) }
 
     unmockkObject(ResourceMapper)
   }
 
   @Test
-  fun testExtractAndSaveResourcesWithExtrcationExtensionAndNullResourceShouldAssignTags() {
-    mockkObject(ResourceMapper)
-
-    coEvery { fhirEngine.load(StructureMap::class.java, any()) } returns StructureMap()
-    coEvery { ResourceMapper.extract(any(), any(), any(), any()) } returns
-      Bundle().apply { addEntry().apply { this.resource = Patient().apply { id = "123456" } } }
-
+  fun testExtractAndSaveResourcesWithExtractionExtensionAndNullResourceShouldAssignTags() {
     val questionnaire =
       Questionnaire().apply {
         addUseContext().apply {
@@ -150,8 +146,14 @@ class QuestionnaireViewModelTest : RobolectricTest() {
           value = CodeableConcept().apply { addCoding().apply { code = "1234567" } }
         }
         addExtension(
-          "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-targetStructureMap",
-          CanonicalType("1234")
+          Extension(
+            "http://hl7.org/fhir/uv/sdc/StructureDefinition/sdc-questionnaire-itemExtractionContext",
+            Expression().apply {
+              language = "application/x-fhir-query"
+              expression = "Patient"
+              name = "Patient"
+            }
+          )
         )
       }
 
@@ -164,12 +166,9 @@ class QuestionnaireViewModelTest : RobolectricTest() {
 
     val patientSlot = slot<Patient>()
 
-    coVerify { ResourceMapper.extract(any(), any(), any(), any()) }
     coVerify { defaultRepo.addOrUpdate(capture(patientSlot)) }
 
     Assert.assertEquals("1234567", patientSlot.captured.meta.tagFirstRep.code)
-
-    unmockkObject(ResourceMapper)
   }
 
   @Test